requirements.txt
pytelegrambotapi
python-dotenv
psycopg2-binary
redis
